Move data service to its own directory
