+++ src/app/search-table/index.ts
@@ -1,2 +1,1 @@
-export * from './data.service';
 export * from './search-table.component';
