+++ src/app/search-table/search-table.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { DataService } from './data.service';
+import { DataService } from '../data-service';
 import { TranslateService } from '../translate';
